# docs/swagger.yaml
basePath: /api
definitions:
  auth.LogoutAllDevicesRequest:
    properties:
      user_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: hanguyenviet772004@gmail.com
    name: Việt Hà
  description: API tài liệu cho hệ thống xác thực
  title: GOLANDWEB API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Đăng nhập vào hệ thống với email và mật khẩu để nhận Access Token
        và Refresh Token.
      parameters:
      - description: Email người dùng
        in: body
        name: email
        required: true
        schema:
          type: string
      - description: Mật khẩu người dùng
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Đăng nhập thành công
          schema:
            additionalProperties: true
            type: object
        "40001":
          description: Lỗi hệ thống
          schema:
            additionalProperties: true
            type: object
        "40004":
          description: Lỗi tạo token
          schema:
            additionalProperties: true
            type: object
        "50001":
          description: Không tìm thấy người dùng
          schema:
            additionalProperties: true
            type: object
        "50003":
          description: Sai mật khẩu
          schema:
            additionalProperties: true
            type: object
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Xóa access token hiện tại (thiết bị đang dùng)
      parameters:
      - description: Access Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Đăng xuất thành công
          schema:
            additionalProperties: true
            type: object
        "40001":
          description: Lỗi khi xóa token
          schema:
            additionalProperties: true
            type: object
      tags:
      - Auth
  /api/v1/auth/logout-all:
    post:
      consumes:
      - application/json
      description: Xóa tất cả access token của người dùng dựa vào user_id
      parameters:
      - description: ID người dùng cần đăng xuất toàn bộ
        in: body
        name: user_id
        required: true
        schema:
          $ref: '#/definitions/auth.LogoutAllDevicesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Đăng xuất thành công
          schema:
            additionalProperties: true
            type: object
        "20003":
          description: User ID không hợp lệ
          schema:
            additionalProperties: true
            type: object
        "40001":
          description: Lỗi khi xóa tất cả token
          schema:
            additionalProperties: true
            type: object
      tags:
      - Auth
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Sử dụng refreshToken để lấy accessToken mới.
      parameters:
      - description: Refresh Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token mới
          schema:
            additionalProperties: true
            type: object
        "30002":
          description: Token không hợp lệ
          schema:
            additionalProperties: true
            type: object
        "40002":
          description: Lỗi khi truy vấn Refresh Token
          schema:
            additionalProperties: true
            type: object
        "40003":
          description: Không thể tạo Access Token
          schema:
            additionalProperties: true
            type: object
      tags:
      - Auth
...
swagger: "2.0"